% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npm-assess.R
\name{A_scorer}
\alias{A_scorer}
\title{A score function}
\usage{
A_scorer(energy_score, sugar_score, fat_score, sodium_score)
}
\arguments{
\item{energy_score, }{numeric value for energy score}

\item{sugar_score, }{numeric value for sugar score}

\item{fat_score, }{numeric value for fat score}

\item{sodium_score, }{numeric value for sodium score}
}
\value{
a numeric value of the A score
}
\description{
A scores are defined as the scores for
energy, sugars, fat and sodium summed together
}
