% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{AutoDefaultParamCheck}
\alias{AutoDefaultParamCheck}
\title{Required parameter with automatic defaults checking function}
\usage{
AutoDefaultParamCheck(data_frame, report_data)
}
\arguments{
\item{data_frame}{a data.frame object with product data}

\item{report_data}{a data.frame object produced by \code{ReqParamCheck()} with information on missing parameters}
}
\value{
report_data a data.frame object updated with information on missing parameters
}
\description{
This function checks whether values for \code{drink_format}, \code{drink_type}, or
\code{food_type} are provided. These have default settings which are automatically
applied if a value is not supplied. Please read the documentation for further detail.
To record use of default parameters, please see \code{AutoDefaultParamNote()}.
}
\seealso{
\code{\link[=ReqParamCheck]{ReqParamCheck()}},\code{\link[=AutoDefaultParamNote]{AutoDefaultParamNote()}},\code{\link[=ManualParamUpdate]{ManualParamUpdate()}}
}
