% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{AutoDefaultParamNote}
\alias{AutoDefaultParamNote}
\title{Required parameter with automatic defaults recording function}
\usage{
AutoDefaultParamNote(data_frame, report_data)
}
\arguments{
\item{data_frame}{a data.frame object with product data}

\item{report_data}{a data.frame object produced by \code{ReqParamCheck()} with information on missing parameters}
}
\value{
data_frame the product data with an additional column recording default use
}
\description{
This function checks whether values for \code{drink_format}, \code{drink_type}, or
\code{food_type} are provided and records in the product data frame if defaults
will be used for analysis.
}
\seealso{
\code{\link[=AutoDefaultParamCheck]{AutoDefaultParamCheck()}},\code{\link[=ReqParamCheck]{ReqParamCheck()}}
}
