% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npm-assess.R
\name{C_scorer}
\alias{C_scorer}
\title{C score function}
\usage{
C_scorer(fvn_score, protein_score, fibre_score)
}
\arguments{
\item{fvn_score, }{numeric value for fruit/vegetables/nuts percentage score}

\item{protein_score, }{numeric value for protein score}

\item{fibre_score, }{numeric value for fibre score}
}
\value{
a numeric value of the C score
}
\description{
C score is defined as the scores for fruit/veg/nut percentage,
protein and fibre summed together
}
