% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdrcdrinks}
\alias{cdrcdrinks}
\title{Example dataset called CDRC Drinks}
\format{
\code{cdrcdrinks} ##
A data frame of 10 rows and 21 columns:
\describe{
\item{name}{Name of product}
\item{brand}{Brand of product}
\item{product_category}{Category of product}
\item{product_type}{Type of product, i.e. Food or Drink}
\item{food_type}{Type of food, used for specific gravity conversion}
\item{drink_format}{Format of drink, used for specific gravity conversion}
\item{drink_type}{Type of drink, used for speciifc gravity conversion}
\item{nutrition_info}{How nutritional information is presented, for specific gravity conversion}
\item{energy_measurement_kj}{Energy measurement in KJ}
\item{energy_measurement_kcal}{Energy measurement in kcal}
\item{sugar_measurement_g}{Sugar content measured in in grams}
\item{satfat_measurement_g}{The saturated fat content measured in grams}
\item{salt_measurement_g}{The salt content measured in grams}
\item{sodium_measurement_mg}{The salt content measured in mg of sodium}
\item{fibre_measurement_nsp}{The fibtre measurement using NSP}
\item{fibre_measurement_aoac}{The fibre measurement using AOAC}
\item{protein_measurement_g}{The protein emasurement of the product}
\item{fvn_measurement_percent}{The percent of fruit/veg/nuts in the product}
\item{weight_g}{The mass of the product if solid}
\item{volume_ml}{The volume of the product if liquid}
\item{volume_water_ml}{The volume of water for diluted products}
}
}
\usage{
cdrcdrinks
}
\description{
A handcrafted set of data that can be used as the input for
nutrientprofiler functions
}
\keyword{datasets}
