% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{CheckValues}
\alias{CheckValues}
\title{Check parameter values function}
\usage{
CheckValues(data_frame)
}
\arguments{
\item{data_frame}{a data.frame object with product data}
}
\value{
results a data.frame object with each named parameter and the unique occurrences of values
}
\description{
A function that takes an input dataframe and counts unique occurrences
for each parameter and lists all included values. This allows you to check that no
values have misspelled or invalid entries.
}
