% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npm-scoring.R
\name{energy_value_adjuster}
\alias{energy_value_adjuster}
\title{Function for adjusting energy information for A score calculation}
\usage{
energy_value_adjuster(value, adjusted_weight, adjuster_type = "kj")
}
\arguments{
\item{value}{a numeric value corresponding to a energy measurement in a food/drink}

\item{adjusted_weight}{a numeric value corresponding to the total
weight of the food/drink after specific gravity adjustment}
}
\value{
a numeric value of adjusted nutritional data
}
\description{
Adjustment is required for calculating scores based on nutritional measurements.
}
