% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter-names.R
\name{fillMissingParameters}
\alias{fillMissingParameters}
\title{Fill the missing parameters function}
\usage{
fillMissingParameters(data_frame, missing_parameters)
}
\arguments{
\item{data_frame}{a data.frame object, loaded from a csv or Excel}

\item{missing_parameters}{a vector object of missing column/parameter names}
}
\value{
data_frame a data.frame object, with new columns added filled with NA
}
\description{
This function creates columns filled with \code{NA} for a provided
vector of missing parameter names.
Please attempt to rename parameters before filling in columns.
}
\seealso{
\code{\link[=inputDataCheck]{inputDataCheck()}}
}
