% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npm-scoring.R
\name{fruit_veg_nut_scorer}
\alias{fruit_veg_nut_scorer}
\title{Fruit, Vegetable and nut scorer}
\usage{
fruit_veg_nut_scorer(value)
}
\arguments{
\item{value}{a numeric value of the percentage of fruit, nuts and vegetables}
}
\value{
a numeric score value
}
\description{
The scoring logic for the percentage of fruit, vegetables and nuts
in a product is different to other scorers. Therefore, we can't use
the generic scoring function and have to write a specific one here.
}
