% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npm-scoring.R
\name{generic_adjuster}
\alias{generic_adjuster}
\title{Function for generically adjusting nutritional measurement by total weight}
\usage{
generic_adjuster(value, adjusted_weight)
}
\arguments{
\item{value}{a numeric value corresponding to a nutritional measurement in a food/drink}

\item{adjusted_weight}{a numeric value corresponding to the total
weight of the food/drink after specific gravity adjustment}
}
\value{
a numeric value of adjusted nutritional data
}
\description{
Adjustment is required for calculating scores based on nutritional measurements.
}
