% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific-gravity.R
\name{generic_specific_gravity}
\alias{generic_specific_gravity}
\title{Generic specific gravity converter}
\usage{
generic_specific_gravity(volume, specific_gravity)
}
\arguments{
\item{volume, }{a value representing the volume to adjust by
a specific gravity}

\item{specific_gravity, }{a multiplier representing the specific
gravity of the product}
}
\value{
a value adjusted by the specific_gravity
}
\description{
This function specifies the generic specific gravity conversion
}
