% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter-names.R
\name{inputDataCheck}
\alias{inputDataCheck}
\title{Data check function}
\usage{
inputDataCheck(data_frame)
}
\arguments{
\item{data_frame}{a data.frame object, loaded from a csv or Excel}
}
\value{
Prints code snippets to help you rename variables if needed
}
\description{
This function checks that all required parameter names or
spreadsheet "column" names are present, and provides
a code snippet to change any parameter names required.
}
\seealso{
\code{\link[=parameterRename]{parameterRename()}}
}
