% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter-names.R
\name{listMissingParameters}
\alias{listMissingParameters}
\title{List the missing parameters function}
\usage{
listMissingParameters(data_frame)
}
\arguments{
\item{data_frame}{a data.frame object, loaded from a csv or Excel}
}
\value{
missing_column_names a vector object with missing column names
}
\description{
This function returns the parameters that your input data file is missing.
}
\seealso{
\code{\link[=inputDataCheck]{inputDataCheck()}}
}
