% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{ManualParamUpdate}
\alias{ManualParamUpdate}
\title{Manual parameter update function}
\usage{
ManualParamUpdate(data_frame, parameter_name, index_list, value)
}
\arguments{
\item{data_frame}{a data.frame object with product data}

\item{parameter_name}{a chr object with name of parameter to be changed}

\item{index_list}{a vector object of index values of parameters to be changed}

\item{value}{the value to enter for the parameter}
}
\value{
data_frame the product data with an additional column recording manual param editing
}
\description{
This function allows updating/setting of a parameter manually where one is missing or
incorrect. Additionally, it records which parameters have been manually added. This function
allows multiple rows of data to be changed at once by supplying a list of index values.
}
\seealso{
\code{\link[=AutoDefaultParamCheck]{AutoDefaultParamCheck()}},\code{\link[=ReqParamCheck]{ReqParamCheck()}}
}
