% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npm-assess.R
\name{NPM_total}
\alias{NPM_total}
\title{NPM_total}
\usage{
NPM_total(a_score, c_score, fvn_score, fibre_score)
}
\arguments{
\item{a_score, }{a numeric value for the A score}

\item{c_score, }{a numeric value for the C score}

\item{fvn_score, }{a numeric value for the specific score for fruit/veg/nuts percentage}

\item{fibre_score, }{a numeric value for the fibre score}
}
\value{
a numeric value for the overall NPM score
}
\description{
A function to calculate total NPM score
Logic for this function is documented in
\href{https://assets.publishing.service.gov.uk/government/uploads/system/uploads/attachment_data/file/216094/dh_123492.pdf}{Nutrient Profiling Technical Guidance}
page 6.
}
