% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npm-scoring.R
\name{NPMScore}
\alias{NPMScore}
\title{NPM scoring dispatcher}
\usage{
NPMScore(row, sg_adjusted_label)
}
\arguments{
\item{row}{the row from an NPMCalculator dataframe}

\item{sg_adjusted_label}{a character value specifying the name of the specific gravity adjusted column}
}
\value{
a matrix of NPM scores
}
\description{
This function is written for calculating NPM scores across an entire row
in a dataframe. It is wrapper logic for the underlying NPM_score_function
when applied to an NPMCalculator input data frame as such it is highly inflexible without
expected column names. It expects the specific gravity conversions to already have been performed
and values outputted to a new single column.
}
