% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter-names.R
\name{parameterRename}
\alias{parameterRename}
\title{Parameter rename function}
\usage{
parameterRename(missing_column_name, associated_data_column, data_frame)
}
\arguments{
\item{missing_column_name}{a chr object with the missing nutrientprofiler parameter}

\item{associated_data_column}{a chr object with the associated column header from the loaded data}

\item{data_frame}{a data.frame object, loaded from a csv or Excel}
}
\value{
data_frame a data.frame object with updated column names
}
\description{
This function renames specified parameters/columns in your dataframe
to match the expected parameters required by nutrientprofiler.
}
\seealso{
\code{\link[=inputDataCheck]{inputDataCheck()}}
}
