% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{ReqParamCheck}
\alias{ReqParamCheck}
\title{Required parameter checking function}
\usage{
ReqParamCheck(data_frame)
}
\arguments{
\item{data_frame}{a data.frame object}
}
\value{
report_data a data.frame object with information on missing parameters
}
\description{
This function processes the data.frame object with product data
and check for required parameters. It returns a report table which
contains row indices of missing required parameters. Please
see guidance on required and default parameters for more information
and before processing your data.
}
\seealso{
\code{\link[=SaveReport]{SaveReport()}},\code{\link[=AutoDefaultParamCheck]{AutoDefaultParamCheck()}},\code{\link[=AutoDefaultParamNote]{AutoDefaultParamNote()}},\code{\link[=ManualParamUpdate]{ManualParamUpdate()}}
}
