% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npm-scoring.R
\name{salt_adjuster}
\alias{salt_adjuster}
\title{Function for adjusting the salt input value}
\usage{
salt_adjuster(value, adjusted_weight, adjuster_type = "sodium")
}
\arguments{
\item{value}{a numeric value corresponding to a salt measurement in a food/drink}

\item{adjusted_weight}{a numeric value corresponding to the total
weight of the food/drink after specific gravity adjustment}

\item{type}{a character of either "salt" or "sodium" to help determine the required adjustment}
}
\value{
a numeric value with appropriate adjustment made
}
\description{
Adjustments is required for calculating scores and depends on the type of salt measurement provided.
}
