% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{SaveReport}
\alias{SaveReport}
\title{Save report function}
\usage{
SaveReport(data_frame, file_path)
}
\arguments{
\item{data_frame}{a data.frame object}

\item{file_path}{a chr object containing relative filepath from the current working directory, to save csv}
}
\value{
saves csv file
}
\description{
This function saves a transposed version of the report generated by \code{ReqParamCheck()}
}
\seealso{
\code{\link[=ReqParamCheck]{ReqParamCheck()}},\code{\link[=AutoDefaultParamCheck]{AutoDefaultParamCheck()}},\code{\link[=AutoDefaultParamNote]{AutoDefaultParamNote()}},\code{\link[=ManualParamUpdate]{ManualParamUpdate()}}
}
