% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npm-scoring.R
\name{scoring_function}
\alias{scoring_function}
\title{a template scoring function}
\usage{
scoring_function(value, thresholds)
}
\arguments{
\item{value}{a passed numeric value}

\item{thresholds}{a vector of thresholds to use to score against in order of highest to lowest}
}
\description{
This function takes a value and a vector of thresholds
It iterates over the reversed vector of indexes in the thresholds vector
And checks if the value is less than the threshold value at the itered index
(comparing from smallest to largest).
The threshold score is calculated as the length of the thresholds minus every else branch iter
of the loop.
}
