% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific-gravity.R
\name{sg_cord_drink_converter}
\alias{sg_cord_drink_converter}
\title{Specific gravity drink converter for cordial drinks}
\usage{
sg_cord_drink_converter(row)
}
\arguments{
\item{row}{a row from a data.frame object}
}
\value{
a numeric value of the specific gravity adjusted volume depending on \code{nutrition_info} column.
}
\description{
This function performs a specific gravity conversion for cordial drinks
the conversion performed is based on the value in the \code{nutrition_info} column.
}
