% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific-gravity.R
\name{sg_powd_drink_converter}
\alias{sg_powd_drink_converter}
\title{Specific gravity drink converter for powdered drinks}
\usage{
sg_powd_drink_converter(row)
}
\arguments{
\item{row}{a row from a data.frame object}
}
\value{
a numeric value of the specific gravity adjusted volume/mass
depending on \code{nutrition_info} column.
}
\description{
This function performs a specific gravity conversion for powdered drinks
the conversion performed is based on the value in the \code{nutrition_info}
column.
}
