% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific-gravity.R
\name{SGConverter}
\alias{SGConverter}
\title{Specific Gravity converter function}
\usage{
SGConverter(row)
}
\arguments{
\item{row}{a row from a data.frame object}
}
\value{
A number.
}
\description{
This function is the key entry point for performing specific gravity conversions
It works by processing data from a dataframe row and dispatching based on column values
to a series of additional sub functions
These return back an SG-adjusted number
}
