% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific-gravity.R
\docType{data}
\name{SGtab}
\alias{SGtab}
\title{specific gravity conversion named vector}
\format{
An object of class \code{numeric} of length 14.
}
\usage{
SGtab
}
\description{
specific gravity conversion named vector
}
\keyword{datasets}
