% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npm-assess.R
\name{A_scorer}
\alias{A_scorer}
\title{A score function}
\usage{
A_scorer(energy_score, sugar_score, satfat_score, salt_score)
}
\arguments{
\item{energy_score, }{numeric value for energy score}

\item{sugar_score, }{numeric value for sugar score}

\item{satfat_score, }{numeric value for sat fat score}

\item{salt_score, }{numeric value for salt score}
}
\value{
a numeric value of the A score
}
\description{
A scores are defined as the scores for
energy, sugars, sat fat and salt summed together
}
