% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific-gravity.R
\name{sg_food_converter}
\alias{sg_food_converter}
\title{Specific gravity food dispatcher}
\usage{
sg_food_converter(row)
}
\description{
This function is run for data with the \code{product_type} "food"
It checks if the column \code{weight_g} contains NA values to assume whether the food is liquid or solid
And if solid returns the weight unadjusted or if liquid dispatches to an additional function
}
